pyproject.toml
[project]
name = "flax-js"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chz>=0.4.0",
    "huggingface-hub>=1.4.1",
    "ipykernel>=7.2.0",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
    "safetensors>=0.7.0",
    "torch>=2.10.0",
    "transformers>=5.1.0",
]
